+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Added
 - "payment" account type for Pix related resources
 - Event.WorkspaceId property to allow multiple Workspace Webhook identification
+- Workspace.Update() to allow parameter updates
+- AllowedTaxIds property to Workspace resource
 
 ### Fixed
 - special characters in BrcodePreview query